fix isValid v1.0
