+++ .env.example
@@ -1,3 +1,2 @@
 HYGRAPH_GRAPHQL_URI=
-HYGRAPH_POSTS_WEBHOOK_KEY=
-HYGRAPH_BLOGS_WEBHOOK_KEY=
+HYGRAPH_WEBHOOK_SECRET_REVALIDATE=
